backend/models/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.1
pydantic==2.5.2

# ML/AI Core
numpy==1.24.3
torch>=2.0.0
scikit-learn==1.3.2

# NLP (for chatbot)
transformers==4.36.0
sentence-transformers==2.2.2
langchain==0.0.348

# Computer Vision (for pose)
mediapipe==0.10.8
opencv-python-headless==4.8.1.78
Pillow==10.1.0

# Utilities
aiofiles==23.2.1
